+++ downloader/Objective-Zip/ZipReadStream.m
@@ -57,7 +57,7 @@
 }
 
 - (int)readDataWithBuffer:(NSMutableData *)buffer {
-	int err = unzReadCurrentFile(_unzFile, buffer.mutableBytes, buffer.length);
+	int err = (int)unzReadCurrentFile(_unzFile, buffer.mutableBytes, (int)buffer.length);
 	if (err < 0) {
 		NSString *reason = [NSString stringWithFormat:@"Error in reading \"%@\" in the zipfile",_filenameInZip];
 		@throw [ZipException exceptionWithReason:reason andError:err];
