+++ downloader/Objective-Zip/ZipWriteStream.m
@@ -56,7 +56,7 @@
 }
 
 - (void)writeData:(NSData *)data {
-	int err = zipWriteInFileInZip(_zipFile, data.bytes, data.length);
+	int err = zipWriteInFileInZip(_zipFile, data.bytes, (int)data.length);
 	if (err < 0) {
 		NSString *reason = [NSString stringWithFormat:@"Error writing \"%@\" in the zipfile",_filenameInZip];
 		@throw [ZipException exceptionWithReason:reason andError:err];
